+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.3.0] - 2022-05-17
 ### Added
 - Add `zone list` command that lists available zones.
 - Add `--wait` flag to `server create` and `server stop` commands to wait until server is in `started` and `stopped` state, respectively.
@@ -102,7 +104,7 @@ Initial public beta release :tada:
 ### Added
 - Current feature set added! First internal release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v1.3.0...HEAD
 [0.1.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v0.1.0
 [0.1.1]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v0.1.1
 [0.5.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v0.5.0
@@ -113,3 +115,4 @@ Initial public beta release :tada:
 [1.1.2]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.1.2
 [1.1.3]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.1.3
 [1.2.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.2.0
+[1.3.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.3.0